refactor: canonicalize json contracts
