+++ polylogue/cli/query_output.py
@@ -35,9 +35,9 @@ from polylogue.cli.query_stats import (
     output_stats_by_summaries,
     output_stats_sql,
 )
+from polylogue.lib.json import JSONDocument
 from polylogue.logging import get_logger
 from polylogue.rendering.formatting import format_conversation
-from polylogue.schemas.json_types import JSONDocument
 from polylogue.surface_payloads import ConversationListRowPayload
 
 logger = get_logger(__name__)
