+++ polylogue/schemas/audit_models.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
+from polylogue.lib.json import JSONDocument, json_document
 from polylogue.lib.outcomes import OutcomeCheck, OutcomeReport, OutcomeStatus
-from polylogue.schemas.json_types import JSONDocument, json_document
 
 
 @dataclass
