+++ polylogue/schemas/audit_walkers.py
@@ -7,7 +7,7 @@ from collections.abc import Iterable
 from pathlib import Path
 from typing import TypeAlias
 
-from polylogue.schemas.json_types import JSONDocument, JSONValue, json_document, json_document_list
+from polylogue.lib.json import JSONDocument, JSONValue, json_document, json_document_list
 from polylogue.schemas.runtime_registry import SchemaRegistry
 
 SchemaPath = str
