+++ polylogue/schemas/generation_support.py
@@ -5,12 +5,12 @@ from __future__ import annotations
 from collections.abc import Mapping, Sequence
 from typing import Protocol, TypeAlias, overload
 
+from polylogue.lib.json import JSONDocument, json_document
 from polylogue.schemas import generation_dynamic_keys as _dynamic_keys
 from polylogue.schemas.field_stats import FieldStats
 from polylogue.schemas.generation_field_annotations import annotate_schema, remove_nested_required
 from polylogue.schemas.generation_redaction import _build_redaction_report
 from polylogue.schemas.generation_semantic_relations import annotate_semantic_and_relational
-from polylogue.schemas.json_types import JSONDocument, json_document
 
 SchemaPayload: TypeAlias = JSONDocument
 SchemaMapping: TypeAlias = Mapping[str, object]
