+++ polylogue/schemas/packages.py
@@ -6,7 +6,7 @@ from dataclasses import asdict, dataclass, field
 from datetime import datetime, timezone
 from typing import Literal, TypeAlias
 
-from polylogue.schemas.json_types import JSONDocument, json_document, json_document_list
+from polylogue.lib.json import JSONDocument, json_document, json_document_list
 from polylogue.types import Provider
