+++ polylogue/schemas/pinning.py
@@ -23,7 +23,7 @@ from dataclasses import asdict, dataclass, field
 from pathlib import Path
 from typing import Literal, TypeAlias, cast
 
-from polylogue.schemas.json_types import JSONDocument, json_document, json_document_list
+from polylogue.lib.json import JSONDocument, json_document, json_document_list
 from polylogue.types import Provider
 
 logger = logging.getLogger(__name__)
