+++ polylogue/schemas/runtime_registry.py
@@ -10,11 +10,11 @@ from dataclasses import dataclass
 from datetime import datetime, timezone
 from pathlib import Path
 
+from polylogue.lib.json import json_document
 from polylogue.lib.provider_identity import canonical_schema_provider as _canonical_schema_provider
 from polylogue.lib.provider_identity import normalize_provider_token
 from polylogue.lib.raw_payload_decode import JSONRecord
 from polylogue.paths import data_home
-from polylogue.schemas.json_types import json_document
 from polylogue.schemas.observation import (
     derive_bundle_scope,
     extract_schema_units_from_payload,
