+++ polylogue/schemas/sampling.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 from collections.abc import Iterator
 from pathlib import Path
 
+from polylogue.lib.json import JSONDocument
 from polylogue.lib.raw_payload import build_raw_payload_envelope, collect_limited_samples
 from polylogue.paths import db_path as archive_db_path
-from polylogue.schemas.json_types import JSONDocument
 from polylogue.schemas.observation import resolve_provider_config
 from polylogue.schemas.observation_models import SchemaUnit
 from polylogue.schemas.sampling_db import (
