+++ polylogue/schemas/tooling_registry.py
@@ -8,7 +8,7 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import TYPE_CHECKING, TypeAlias, cast
 
-from polylogue.schemas.json_types import json_document, json_document_list
+from polylogue.lib.json import json_document, json_document_list
 from polylogue.schemas.observation import schema_cluster_id
 from polylogue.schemas.observation_models import SchemaClusterPayload
 from polylogue.schemas.packages import SchemaPackageCatalog, SchemaVersionPackage
