+++ tests/unit/core/test_schema_annotation_contracts.py
@@ -21,6 +21,7 @@ from collections.abc import Mapping
 
 import pytest
 
+from polylogue.lib.json import JSONDocument
 from polylogue.schemas.field_stats import FieldStats, _collect_field_stats
 from polylogue.schemas.generation_support import (
     _annotate_semantic_and_relational,
@@ -28,7 +29,6 @@ from polylogue.schemas.generation_support import (
 from polylogue.schemas.generation_workflow import (
     generate_schema_from_samples,
 )
-from polylogue.schemas.json_types import JSONDocument
 from polylogue.schemas.semantic_inference_runtime import (
     infer_semantic_roles,
 )
