+++ tests/unit/pipeline/test_prepare_semantic.py
@@ -11,10 +11,10 @@ Covers:
 
 from __future__ import annotations
 
+from polylogue.lib.json import JSONDocument, json_document
 from polylogue.lib.viewports import ToolCategory, classify_tool
 from polylogue.pipeline.semantic_capture import extract_subagent_spawns, parse_git_operation
 from polylogue.pipeline.semantic_metadata import extract_tool_metadata
-from polylogue.schemas.json_types import JSONDocument, json_document
 
 
 def _payload(data: object) -> JSONDocument:
